refactor: use Int for money types to fix floating point precision
- Convert all Double money fields (price, amount, balance) to Int cents.
- Update database schema version from 1 to 2.
- Update parsing, mappers, UI string formatting, and cart pricing calculator.
- Adjust seed data and unit tests to reflect cent amounts.

+++ app/src/main/java/com/fakeshopee/app/domain/model/Product.kt
@@ -4,7 +4,7 @@ data class Product(
     val id: String,
     val title: String,
     val category: String,
-    val price: Double,
+    val price: Int,
     val rating: Double,
     val reviewCount: Int,
     val description: String,
@@ -44,6 +44,6 @@ data class CartItem(
 data class Coupon(
     val code: String,
     val discountPercentage: Int = 0,
-    val discountFlat: Double = 0.0,
+    val discountFlat: Int = 0,
     val description: String
 )
